REFACTOR : getData func create & code refactoring
